changed db query, added function to InvState
